CLI-747: Isolate test fixtures (#936)
* CLI-747: Isolate test fixtures

* separate base class

* convenience method

* deprecate legacy format

* Revert "deprecate legacy format"

This reverts commit 172dd503334a8819d4fb33d403364c5b69d52cda.
